+++ app/scripts/services/dimSettingsService.factory.js
@@ -21,6 +21,7 @@
         hideFilteredItems: false,
         itemDetails: false,
         itemStat: false,
+        itemQuality: false,
         showElements: false,
         condensed: false,
         characterOrder: 'mostRecent',